test(nfs): split NfsTasksTest into separate task-specific test classes
